refactor: change package names in dash/
