+++ dash/quorum/selectpeers/select_validators.go
@@ -1,4 +1,6 @@
-package dip6
+// selectpeers package contains algorithm that selects peers based on the deterministic connection selection algorithm
+// described in DIP-6
+package selectpeers
 
 import (
 	"fmt"
