+++ dash/quorum/selectpeers/select_validators_test.go
@@ -1,10 +1,10 @@
-package dip6
+package selectpeers
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tendermint/tendermint/dash/mock"
+	"github.com/tendermint/tendermint/dash/quorum/mock"
 	"github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/types"
 )
