+++ dash/quorum/selectpeers/sortable_validator.go
@@ -1,4 +1,4 @@
-package dip6
+package selectpeers
 
 import (
 	"bytes"
