+++ dash/quorum/selectpeers/sorted_validator_list.go
@@ -1,4 +1,4 @@
-package dip6
+package selectpeers
 
 import (
 	"bytes"
